+++ packages/admin/src/app/feature/fact-types/fact-types.store.ts
@@ -30,6 +30,7 @@ interface FactTypesState {
   pageIndex: number;
   sort: string;
   order: SortDirection;
+  search: { field: string; searchString: string }[];
 }
 
 const initialState: FactTypesState = {
@@ -40,6 +41,7 @@ const initialState: FactTypesState = {
   pageIndex: 0,
   sort: 'name',
   order: 'asc',
+  search: [],
 };
 
 export const FactTypesStore = signalStore(
@@ -58,6 +60,7 @@ export const FactTypesStore = signalStore(
         sort?: string;
         order?: SortDirection;
         pageIndex?: number;
+        search?: { field: string; searchString: string }[];
       }) {
         patchState(
           store,
@@ -98,6 +101,7 @@ export const FactTypesStore = signalStore(
           factTypesService.getFactTypes({
             limit: store.limit(),
             offset: store.offset(),
+            search: store.search(),
           }),
         );