+++ packages/server/app/helpers/query_builder.ts
@@ -6,6 +6,19 @@ import Resource from '#models/resource'
 import Person from '#models/person'
 import { TransactionClientContract } from '@adonisjs/lucid/types/database'
 import Tag from '#models/tag'
+import { getIdsByDimensionQuery } from '#helpers/query_dimensions'
+
+const defaultSearchFields: Record<string, string[]> = {
+  cohorts: ['name'],
+  event_types: ['name'],
+  fact_types: ['name'],
+  people: ['family_name', 'given_name'],
+  person_types: ['name'],
+  resources: ['name'],
+  resource_types: ['name'],
+  tags: ['pathname'],
+  users: ['email'],
+}
 
 function getSortSettings(queryData: Record<string, any> = {}) {
   const sort = queryData['sort']
@@ -14,16 +27,28 @@ function getSortSettings(queryData: Record<string, any> = {}) {
   return [sort, order]
 }
 
-export function buildApiQuery(
-  modelQuery: any,
-  queryData: Record<string, any> = { limit: 10, offset: 0 },
-  tableName: string,
+export async function buildApiQuery({
+  modelQuery,
+  queryData = { limit: 10, offset: 0 },
+  tableName,
+}: {
+  modelQuery: any
+  queryData?: Record<string, any>
+  tableName: string
   defaultSearch?: string
-) {
+}) {
   let countQuery = db.from(tableName)
   const limit = queryData['limit']
   const offset = queryData['offset']
-  const search = queryData['search']
+  const searchQuery = queryData['search']
+  const filters = queryData['filter']
+  const dims = queryData['dim']
+  if (filters || dims) {
+    const typeTableName = string.singular(tableName) + '_types'
+    const result = await getIdsByDimensionQuery(tableName, typeTableName, filters, dims)
+    modelQuery.andWhereIn('id', result)
+    countQuery.andWhereIn('id', result)
+  }
 
   if (limit) {
     modelQuery.limit(limit)
@@ -32,18 +57,40 @@ export function buildApiQuery(
     modelQuery.offset(offset)
   }
 
-  // Add search functionality to modelQuery
-  if (typeof search === 'string' && defaultSearch) {
-    modelQuery.whereILike(defaultSearch, search)
-    countQuery.whereILike(defaultSearch, search)
-  } else if (typeof search === 'object' && search !== null) {
-    for (const key in search) {
-      if (search.hasOwnProperty(key)) {
-        const searchString = '%' + search[key] + '%'
-        modelQuery.whereILike(string.camelCase(key), searchString)
-        countQuery.whereILike(string.snakeCase(key), searchString)
-      }
+  // Searching not supported for Events or Facts. Just ignore for those.
+  if (tableName !== 'events' && tableName !== 'facts') {
+    let search: any
+    // If simple string passed as search param, convert to search object filter so every
+    // request uses the same search code mechanism below
+    if (typeof searchQuery === 'string') {
+      search = {}
+      defaultSearchFields[tableName].forEach((key: string) => {
+        search[key] = searchQuery
+      })
+    } else {
+      search = searchQuery
     }
+
+    modelQuery.where((query: any) => {
+      // Wrap all of this in .where() so the OR clauses generated below will all be enclosed in
+      // parentheses in the generated SQL as one logical unit.
+      for (const key in search) {
+        if (search.hasOwnProperty(key)) {
+          const searchString = '%' + search[key] + '%'
+          // Specify table name with key to avoid ambiguous column reference error when combining
+          // query with "types" table. Use OR clause so a match in any specified column will be returned.
+          query.orWhere((subQuery: any) => subQuery.whereILike(`${tableName}.${key}`, searchString))
+        }
+      }
+    })
+    countQuery.where((query: any) => {
+      for (const key in search) {
+        if (search.hasOwnProperty(key)) {
+          const searchString = '%' + search[key] + '%'
+          query.orWhere((subQuery: any) => subQuery.whereILike(`${tableName}.${key}`, searchString))
+        }
+      }
+    })
   }
 
   return [modelQuery, countQuery]
@@ -64,6 +111,7 @@ export function buildEventsSort(
         eventsQuery
           .join('event_types', 'event_types.key', 'events.type_key')
           .orderBy('event_types.name', order)
+          .select('events.*')
         break
       case 'person':
         eventsQuery
@@ -98,6 +146,7 @@ export function buildFactsSort(
         factsQuery
           .join('fact_types', 'fact_types.key', 'facts.type_key')
           .orderBy('fact_types.name', order)
+          .select('facts.*')
         break
       case 'person':
         factsQuery
@@ -135,6 +184,7 @@ export function buildPeopleSort(
     const camelSortStr = string.camelCase(sort)
     if (camelSortStr === 'personType') {
       peopleQuery
+        .select('people.*')
         .join('person_types', 'person_types.key', 'people.type_key')
         .orderBy('person_types.name', order)
     } else {
@@ -156,13 +206,16 @@ export function buildResourcesSort(
     const camelSortStr = string.camelCase(sort)
     if (camelSortStr === 'resourceType') {
       resourcesQuery
+        // Need to select the resources columns specifically to avoid aliasing issues with the
+        // resource_types columns
+        .select('resources.*')
         .join('resource_types', 'resource_types.key', 'resources.type_key')
         .orderBy('resource_types.name', order)
     } else {
-      resourcesQuery.orderBy(camelSortStr, order)
+      resourcesQuery.orderBy(`resources.${camelSortStr}`, order)
     }
   } else {
-    resourcesQuery.orderBy('name', 'asc')
+    resourcesQuery.orderBy('resources.name', 'asc')
   }
 }