+++ CHANGELOG.md
@@ -8,16 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Pending]
 
 ### Added
-
-- 
+- All list screens except Events and Facts can be filtered by search terms in Admin
+- Filter and dimension query logic added to all appropriate objects
 
 ### Changed
 
-- 
+- All searches use the object filter system
 
 ### Removed
 
 - 
+
+### Fixed
+
+- Sorting bug on Resources list
 
 ## [0.1.0] - 2024-10-21